backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.110.1
uvicorn==0.25.0
python-dotenv==1.0.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.6.4
pydantic-settings==2.2.1

# Database
motor==3.3.1
pymongo==4.5.0

# Security & Authentication
cryptography==42.0.8
pyjwt==2.10.1
bcrypt==4.1.2

# HTTP Client
requests==2.31.0
aiohttp==3.9.3
httpx==0.27.0

# AI/ML Dependencies
openai==1.35.0
anthropic==0.34.0
tensorflow==2.15.0
scikit-learn==1.4.0
numpy==1.26.0
pandas==2.2.0
torch==2.2.0
transformers==4.37.0

# Blockchain
web3==6.15.1
eth-account==0.10.0
py-solc-x==1.1.1

# Post-Quantum Cryptography - Certified Libraries
# Using pqcrypto for NIST-approved algorithms
pqcrypto==0.3.4
liboqs-python==0.14.0
pycryptodome==3.20.0

# Monitoring & Logging
python-json-logger==2.0.7
structlog==23.2.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.0
pytest-cov==4.0.0
httpx==0.27.0

# Code Quality
black==24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0

# Utils
typer==0.9.0
rich==13.7.0
pydantic-settings==2.2.1
email-validator==2.2.0